COMMERCE-7217 Moved shop by diagram web module and update references
