+++ modules/dxp/apps/commerce/commerce-shop-by-diagram-web/bnd.bnd
