+++ modules/dxp/apps/commerce/commerce-shop-by-diagram-web/dev/public/js/static-env-utils.js
