+++ modules/dxp/apps/commerce/commerce-shop-by-diagram-web/src/main/java/com/liferay/commerce/shop/by/diagram/admin/web/internal/upload/CSDiagramSettingImageUploadFileEntryHandler.java
@@ -1,0 +1,162 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.commerce.shop.by.diagram.admin.web.internal.upload;
+
+import com.liferay.commerce.product.exception.CPAttachmentFileEntryNameException;
+import com.liferay.commerce.product.exception.CPAttachmentFileEntrySizeException;
+import com.liferay.commerce.shop.by.diagram.configuration.CSDiagramSettingImageConfiguration;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.kernel.util.FileUtil;
+import com.liferay.portal.kernel.util.TempFileEntryUtil;
+import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.upload.UniqueFileNameProvider;
+import com.liferay.upload.UploadFileEntryHandler;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import java.util.Map;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ConfigurationPolicy;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	configurationPid = "com.liferay.commerce.shop.by.diagram.configuration.CSDiagramSettingImageConfiguration",
+	configurationPolicy = ConfigurationPolicy.OPTIONAL, enabled = false,
+	service = CSDiagramSettingImageUploadFileEntryHandler.class
+)
+public class CSDiagramSettingImageUploadFileEntryHandler
+	implements UploadFileEntryHandler {
+
+	@Override
+	public FileEntry upload(UploadPortletRequest uploadPortletRequest)
+		throws IOException, PortalException {
+
+		String fileName = uploadPortletRequest.getFileName(_PARAMETER_NAME);
+
+		_validateFile(fileName, uploadPortletRequest.getSize(_PARAMETER_NAME));
+
+		String contentType = uploadPortletRequest.getContentType(
+			_PARAMETER_NAME);
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)uploadPortletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		try (InputStream inputStream = uploadPortletRequest.getFileAsStream(
+				_PARAMETER_NAME)) {
+
+			return addFileEntry(
+				fileName, contentType, inputStream, themeDisplay);
+		}
+	}
+
+	@Activate
+	protected void activate(Map<String, Object> properties) {
+		_csDiagramSettingImageConfiguration =
+			ConfigurableUtil.createConfigurable(
+				CSDiagramSettingImageConfiguration.class, properties);
+	}
+
+	protected FileEntry addFileEntry(
+			String fileName, String contentType, InputStream inputStream,
+			ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		String uniqueFileName = _uniqueFileNameProvider.provide(
+			fileName, curFileName -> _exists(themeDisplay, curFileName));
+
+		Company company = themeDisplay.getCompany();
+
+		return TempFileEntryUtil.addTempFileEntry(
+			company.getGroupId(), themeDisplay.getUserId(), _TEMP_FOLDER_NAME,
+			uniqueFileName, inputStream, contentType);
+	}
+
+	private boolean _exists(ThemeDisplay themeDisplay, String curFileName) {
+		try {
+			FileEntry tempFileEntry = TempFileEntryUtil.getTempFileEntry(
+				themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
+				_TEMP_FOLDER_NAME, curFileName);
+
+			if (tempFileEntry != null) {
+				return true;
+			}
+
+			return false;
+		}
+		catch (PortalException portalException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(portalException, portalException);
+			}
+
+			return false;
+		}
+	}
+
+	private void _validateFile(String fileName, long size)
+		throws PortalException {
+
+		if ((_csDiagramSettingImageConfiguration.imageMaxSize() > 0) &&
+			(size > _csDiagramSettingImageConfiguration.imageMaxSize())) {
+
+			throw new CPAttachmentFileEntrySizeException();
+		}
+
+		String extension = FileUtil.getExtension(fileName);
+
+		String[] imageExtensions =
+			_csDiagramSettingImageConfiguration.imageExtensions();
+
+		for (String imageExtension : imageExtensions) {
+			if (StringPool.STAR.equals(imageExtension) ||
+				imageExtension.equals(StringPool.PERIOD + extension)) {
+
+				return;
+			}
+		}
+
+		throw new CPAttachmentFileEntryNameException(
+			"Invalid image for file name " + fileName);
+	}
+
+	private static final String _PARAMETER_NAME = "imageSelectorFileName";
+
+	private static final String _TEMP_FOLDER_NAME =
+		CSDiagramSettingImageUploadFileEntryHandler.class.getName();
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		CSDiagramSettingImageUploadFileEntryHandler.class);
+
+	private volatile CSDiagramSettingImageConfiguration
+		_csDiagramSettingImageConfiguration;
+
+	@Reference
+	private UniqueFileNameProvider _uniqueFileNameProvider;
+
+}
