+++ modules/dxp/apps/commerce/commerce-shop-by-diagram-web/src/main/java/com/liferay/commerce/shop/by/diagram/admin/web/internal/upload/CSDiagramSettingImageUploadResponseHandler.java
@@ -1,0 +1,122 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.commerce.shop.by.diagram.admin.web.internal.upload;
+
+import com.liferay.commerce.product.exception.CPAttachmentFileEntryNameException;
+import com.liferay.commerce.product.exception.CPAttachmentFileEntrySizeException;
+import com.liferay.commerce.shop.by.diagram.configuration.CSDiagramSettingImageConfiguration;
+import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONFactory;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.servlet.ServletResponseConstants;
+import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.upload.UploadResponseHandler;
+
+import java.util.Map;
+
+import javax.portlet.PortletRequest;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ConfigurationPolicy;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	configurationPid = "com.liferay.commerce.shop.by.diagram.configuration.CSDiagramSettingImageConfiguration",
+	configurationPolicy = ConfigurationPolicy.OPTIONAL, enabled = false,
+	service = CSDiagramSettingImageUploadResponseHandler.class
+)
+public class CSDiagramSettingImageUploadResponseHandler
+	implements UploadResponseHandler {
+
+	@Override
+	public JSONObject onFailure(
+			PortletRequest portletRequest, PortalException portalException)
+		throws PortalException {
+
+		JSONObject jsonObject = _itemSelectorUploadResponseHandler.onFailure(
+			portletRequest, portalException);
+
+		if (portalException instanceof CPAttachmentFileEntryNameException ||
+			portalException instanceof CPAttachmentFileEntrySizeException) {
+
+			JSONObject errorJSONObject = _jsonFactory.createJSONObject();
+
+			String errorMessage = StringPool.BLANK;
+			int errorType = 0;
+
+			if (portalException instanceof CPAttachmentFileEntryNameException) {
+				errorMessage = StringUtil.merge(
+					_csDiagramSettingImageConfiguration.imageExtensions());
+
+				errorType =
+					ServletResponseConstants.SC_FILE_EXTENSION_EXCEPTION;
+			}
+			else if (portalException instanceof
+						CPAttachmentFileEntrySizeException) {
+
+				errorType = ServletResponseConstants.SC_FILE_SIZE_EXCEPTION;
+			}
+
+			errorJSONObject.put(
+				"errorType", errorType
+			).put(
+				"message", errorMessage
+			);
+
+			jsonObject.put("error", errorJSONObject);
+		}
+		else {
+			throw portalException;
+		}
+
+		return jsonObject;
+	}
+
+	@Override
+	public JSONObject onSuccess(
+			UploadPortletRequest uploadPortletRequest, FileEntry fileEntry)
+		throws PortalException {
+
+		return _itemSelectorUploadResponseHandler.onSuccess(
+			uploadPortletRequest, fileEntry);
+	}
+
+	@Activate
+	protected void activate(Map<String, Object> properties) {
+		_csDiagramSettingImageConfiguration =
+			ConfigurableUtil.createConfigurable(
+				CSDiagramSettingImageConfiguration.class, properties);
+	}
+
+	private volatile CSDiagramSettingImageConfiguration
+		_csDiagramSettingImageConfiguration;
+
+	@Reference
+	private ItemSelectorUploadResponseHandler
+		_itemSelectorUploadResponseHandler;
+
+	@Reference
+	private JSONFactory _jsonFactory;
+
+}
