+++ modules/dxp/apps/commerce/commerce-shop-by-diagram-web/src/main/resources/META-INF/resources/js/edit_cp_definition_diagram_setting.js
