--- modules/dxp/apps/commerce/commerce-shop-by-diagram/commerce-shop-by-diagram-web/build.gradle
@@ -1,24 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.util.taglib", version: "default"
-	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.1"
-	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	compileOnly project(":apps:commerce:commerce-account-api")
-	compileOnly project(":apps:commerce:commerce-frontend-taglib")
-	compileOnly project(":apps:commerce:commerce-media-api")
-	compileOnly project(":apps:commerce:commerce-product-api")
-	compileOnly project(":apps:commerce:commerce-product-content-api")
-	compileOnly project(":apps:document-library:document-library-api")
-	compileOnly project(":apps:frontend-taglib:frontend-taglib")
-	compileOnly project(":apps:frontend-taglib:frontend-taglib-clay")
-	compileOnly project(":apps:frontend-taglib:frontend-taglib-react")
-	compileOnly project(":apps:item-selector:item-selector-api")
-	compileOnly project(":apps:item-selector:item-selector-criteria-api")
-	compileOnly project(":apps:item-selector:item-selector-taglib")
-	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
-	compileOnly project(":apps:upload:upload-api")
-	compileOnly project(":core:petra:petra-lang")
-	compileOnly project(":core:petra:petra-string")
-	compileOnly project(":dxp:apps:commerce:commerce-shop-by-diagram:commerce-shop-by-diagram-api")
-}
