--- modules/dxp/apps/commerce/commerce-shop-by-diagram/commerce-shop-by-diagram-web/src/main/java/com/liferay/commerce/shop/by/diagram/admin/web/internal/product/type/CSDiagramCPContentRenderer.java
@@ -1,97 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.commerce.shop.by.diagram.admin.web.internal.product.type;
-
-import com.liferay.commerce.product.catalog.CPCatalogEntry;
-import com.liferay.commerce.product.content.render.CPContentRenderer;
-import com.liferay.commerce.shop.by.diagram.admin.web.internal.display.context.CSDiagramCPTypeDisplayContext;
-import com.liferay.commerce.shop.by.diagram.constants.CSDiagramCPTypeConstants;
-import com.liferay.commerce.shop.by.diagram.constants.CSDiagramWebKeys;
-import com.liferay.commerce.shop.by.diagram.service.CPDefinitionDiagramSettingService;
-import com.liferay.document.library.util.DLURLHelper;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
-
-import java.util.Locale;
-import java.util.ResourceBundle;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Alessio Antonio Rendina
- */
-@Component(
-	enabled = false, immediate = true,
-	property = {
-		"commerce.product.content.renderer.key=" + CSDiagramCPTypeConstants.NAME,
-		"commerce.product.content.renderer.order=" + Integer.MIN_VALUE,
-		"commerce.product.content.renderer.type=" + CSDiagramCPTypeConstants.NAME
-	},
-	service = CPContentRenderer.class
-)
-public class CSDiagramCPContentRenderer implements CPContentRenderer {
-
-	@Override
-	public String getKey() {
-		return CSDiagramCPTypeConstants.NAME;
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, getClass());
-
-		return LanguageUtil.get(resourceBundle, CSDiagramCPTypeConstants.NAME);
-	}
-
-	@Override
-	public void render(
-			CPCatalogEntry cpCatalogEntry,
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws Exception {
-
-		httpServletRequest.setAttribute(
-			CSDiagramWebKeys.CS_DIAGRAM_CP_TYPE_DISPLAY_CONTEXT,
-			new CSDiagramCPTypeDisplayContext(
-				_cpDefinitionDiagramSettingService, _dlURLHelper));
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/render/view.jsp");
-	}
-
-	@Reference
-	private CPDefinitionDiagramSettingService
-		_cpDefinitionDiagramSettingService;
-
-	@Reference
-	private DLURLHelper _dlURLHelper;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.shop.by.diagram.web)"
-	)
-	private ServletContext _servletContext;
-
-}
